Images now work
Had to use dynamic image referencing (which is getImageURL) to get images dynamically and render them on the site.

Changed the names of the images to make use of less code to get images.

Now, all that's left is a bit of styling to make the page look good then it's done.